@@ -1,6 +1,6 @@
 {
   "name": "fcl-web-app",
-  "version": "1.1.1",
+  "version": "1.2.0",
   "license": "GPL v3",
   "scripts": {
     "ng": "ng",
@@ -29,26 +29,27 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^8.2.14",
-    "@angular/cdk": "^8.2.3",
-    "@angular/common": "^8.2.14",
-    "@angular/compiler": "^8.2.14",
-    "@angular/core": "^8.2.14",
-    "@angular/forms": "^8.2.14",
-    "@angular/material": "^8.2.3",
-    "@angular/platform-browser": "^8.2.14",
-    "@angular/platform-browser-dynamic": "^8.2.14",
-    "@angular/router": "^8.2.14",
+    "@angular/animations": "^10.2.5",
+    "@angular/cdk": "^10.2.7",
+    "@angular/common": "^10.2.5",
+    "@angular/compiler": "^10.2.5",
+    "@angular/core": "^10.2.5",
+    "@angular/forms": "^10.2.5",
+    "@angular/material": "^10.2.7",
+    "@angular/platform-browser": "^10.2.5",
+    "@angular/platform-browser-dynamic": "^10.2.5",
+    "@angular/router": "^10.2.5",
     "@auth0/angular-jwt": "^4.0.0",
-    "@ngrx/effects": "^8.6.0",
-    "@ngrx/store": "^8.6.0",
-    "@ngrx/store-devtools": "^8.6.0",
-    "@swimlane/ngx-datatable": "^16.1.1",
+    "@ngrx/effects": "^10.1.2",
+    "@ngrx/store": "^10.1.2",
+    "@ngrx/store-devtools": "^10.1.2",
+    "@swimlane/ngx-datatable": "^18.0.0",
+    "@types/seedrandom": "^3.0.0",
     "ajv": "^6.11.0",
     "angular-password-strength-meter": "^2.0.0",
-    "core-js": "^3.6.4",
+    "core-js": "^3.12.1",
     "css-element-queries": "^1.2.3",
-    "cytoscape": "^3.15.2",
+    "cytoscape": "~3.18.2",
     "cytoscape-cola": "^2.3.1",
     "cytoscape-dagre": "^2.2.2",
     "cytoscape-spread": "^3.0.0",
@@ -57,48 +58,50 @@
     "html2canvas": "^1.0.0-rc.5",
     "immutable": "^3.8.2",
     "javascript-lp-solver": "0.4.5",
-    "lodash": "^4.17.19",
+    "lodash": "^4.17.21",
     "material-design-icons": "^3.0.1",
     "moment": "^2.24.0",
     "mxgraph": "^4.1.0",
+    "ngx-color-picker": "^10.1.0",
     "normalize.css": "^8.0.1",
     "ol": "^5.3.3",
-    "rxjs": "^6.5.4",
-    "tslib": "^1.10.0",
-    "zone.js": "^0.10.2",
+    "rxjs": "^6.6.7",
+    "seedrandom": "^3.0.5",
+    "tslib": "^2.2.0",
+    "zone.js": "~0.10.2",
     "zxcvbn": "^4.4.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.25",
-    "@angular/cli": "^8.3.25",
-    "@angular/compiler-cli": "^8.2.14",
-    "@babel/compat-data": "^7.8.0",
+    "@angular-devkit/build-angular": "~0.1002.3",
+    "@angular/cli": "^10.2.3",
+    "@angular/compiler-cli": "^10.2.5",
+    "@babel/compat-data": "^7.14.0",
     "@types/d3": "^5.7.2",
-    "@types/hammerjs": "^2.0.36",
-    "@types/jest": "^25.1.2",
-    "@types/node": "^13.7.1",
+    "@types/hammerjs": "^2.0.39",
+    "@types/jest": "^26.0.23",
+    "@types/node": "^12.20.13",
     "@types/ol": "^5.3.6",
-    "@types/zxcvbn": "^4.4.0",
-    "codelyzer": "^5.2.1",
+    "@types/zxcvbn": "^4.4.1",
+    "codelyzer": "^5.1.2",
     "cypress": "^4.4.0",
-    "jest": "^25.1.0",
-    "jest-canvas-mock": "^2.3.0",
-    "jest-preset-angular": "^8.0.0",
+    "jest": "^26.6.3",
+    "jest-canvas-mock": "^2.3.1",
+    "jest-preset-angular": "^8.4.0",
     "jest-zone-patch": "0.0.10",
-    "ngentest": "^1.4.0",
+    "ngentest": "^1.4.4",
     "node-sass": "^4.14.1",
-    "protractor": "^5.4.2",
+    "protractor": "~7.0.0",
     "rimraf": "^3.0.2",
     "rxjs-tslint": "^0.1.7",
-    "rxjs-tslint-rules": "^4.28.3",
-    "stylelint": "^13.1.0",
+    "rxjs-tslint-rules": "^4.34.8",
+    "stylelint": "^13.13.1",
     "stylelint-config-standard": "^20.0.0",
-    "ts-jest": "^25.2.0",
+    "ts-jest": "^26.5.6",
     "ts-node": "^8.6.2",
-    "tslint": "^6.0.0",
+    "tslint": "~6.1.0",
     "tslint-config-standard": "^9.0.0",
-    "tsutils": "^3.17.1",
-    "typescript": "3.5.3"
+    "tsutils": "^3.21.0",
+    "typescript": "4.0.5"
   },
   "fclConfig": {
     "lastChange": "2019-06-12 11:25:17 +0200"
