@@ -1,11 +1,18 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, ElementRef } from '@angular/core';
+import { OverlayContainer } from '@angular/cdk/overlay';
 
 @Component({
     selector: 'fcl-root',
     templateUrl: './app.component.html'
 })
-export class AppComponent implements OnInit {
+export class AppComponent {
 
-    ngOnInit() {}
-
+    constructor(
+        private overlayContainer: OverlayContainer,
+        private elementRef: ElementRef
+    ) {
+        // Disables all default context menus in webapp
+        (this.elementRef.nativeElement as HTMLElement).oncontextmenu = e => e.preventDefault();
+        this.overlayContainer.getContainerElement().oncontextmenu = e => e.preventDefault();
+    }
 }
