@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { MatButtonModule } from '@angular/material/button';
 import { MatCardModule } from '@angular/material/card';
 import { MatCheckboxModule } from '@angular/material/checkbox';
-import { MAT_LABEL_GLOBAL_OPTIONS } from '@angular/material/core';
+import { MAT_FORM_FIELD_DEFAULT_OPTIONS } from '@angular/material/form-field';
 import { MatDialogModule } from '@angular/material/dialog';
 import { MatDividerModule } from '@angular/material/divider';
 import { MatExpansionModule } from '@angular/material/expansion';
@@ -15,6 +15,7 @@ import { MatRadioModule } from '@angular/material/radio';
 import { MatSelectModule } from '@angular/material/select';
 import { MatSidenavModule } from '@angular/material/sidenav';
 import { MatSliderModule } from '@angular/material/slider';
+import { MatSlideToggleModule } from '@angular/material/slide-toggle';
 import { MatSnackBarModule } from '@angular/material/snack-bar';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { MatTooltipModule } from '@angular/material/tooltip';
@@ -38,6 +39,7 @@ import { MatListModule } from '@angular/material/list';
         MatSelectModule,
         MatSidenavModule,
         MatSliderModule,
+        MatSlideToggleModule,
         MatToolbarModule,
         MatDividerModule,
         MatExpansionModule,
@@ -50,7 +52,7 @@ import { MatListModule } from '@angular/material/list';
         MatListModule
     ],
     providers: [{
-        provide: MAT_LABEL_GLOBAL_OPTIONS,
+        provide: MAT_FORM_FIELD_DEFAULT_OPTIONS,
         useValue: { float: 'auto' }
     }]
 })
