@@ -3,9 +3,9 @@ import { MatSidenav } from '@angular/material/sidenav';
 import { Router } from '@angular/router';
 import * as _ from 'lodash';
 
-import { SchemaGraphComponent } from '../graph/components/schema-graph.component';
+import { SchemaGraphComponent } from '../graph/components/schema-graph/schema-graph.component';
 import { GraphType } from '../data.model';
-import { GisGraphComponent } from '../graph/components/gis-graph.component';
+import { GisGraphComponent } from '../graph/components/gis-graph/gis-graph.component';
 import { environment } from '../../../environments/environment';
 import { MainPageService } from '../../main-page/services/main-page.service';
 import { Store } from '@ngrx/store';
@@ -53,7 +53,6 @@ export class MainTracingComponent implements OnInit, OnDestroy {
         private mainPageService: MainPageService,
         private store: Store<fromTracing.State>
     ) {
-        document.body.oncontextmenu = e => e.preventDefault();
         this.store.dispatch(new tracingActions.TracingActivated({ isActivated: true }));
     }
 
