@@ -1,11 +1,29 @@
-import { StationHighlightingData } from './../../data.model';
+import { BasicGraphState, DataServiceData, DataTable, StationHighlightingRule, StationHighlightingStats, TableColumn } from './../../data.model';
 import { Component, OnInit, OnDestroy } from '@angular/core';
 import { select, Store } from '@ngrx/store';
 import { Observable, Subscription } from 'rxjs';
-import * as fromTracing from '../../state/tracing.reducers';
+import * as tracingReducers from '../../state/tracing.reducers';
+import * as tracingActions from '../../state/tracing.actions';
 import * as tracingSelectors from '../../state/tracing.selectors';
+import * as configurationActions from '../configuration.actions';
 import { takeWhile } from 'rxjs/operators';
 import { AlertService } from '@app/shared/services/alert.service';
+import { HighlightingRuleDeleteRequestData, PropToValuesMap } from '../configuration.model';
+import { DataService } from '@app/tracing/services/data.service';
+import { ComplexFilterUtils } from '../shared/complex-filter-utils';
+import { EditHighlightingService } from '../edit-highlighting.service';
+
+interface HighlightingState {
+    graphState: BasicGraphState;
+    highlightingState: StationHighlightingRule[];
+    editIndex: number | null;
+}
+
+interface CachedData {
+    dataTable: DataTable;
+    propToValuesMap: PropToValuesMap;
+    data: DataServiceData;
+}
 
 @Component({
     selector: 'fcl-highlighting-station',
@@ -14,7 +32,35 @@ import { AlertService } from '@app/shared/services/alert.service';
 })
 export class HighlightingStationComponent implements OnInit, OnDestroy {
 
-    colorAndShapeHighlightings: StationHighlightingData[] = [];
+    get colorOrShapeRuleEditIndex(): number | null {
+        return this.cachedState === null ?
+            null :
+            this.cachedState.editIndex;
+    }
+
+    get rules(): StationHighlightingRule[] {
+        return this.cachedState === null ?
+            [] :
+            this.cachedState.highlightingState;
+    }
+
+    get availableProperties(): TableColumn[] {
+        return this.cachedData ?
+            this.cachedData.dataTable.columns :
+            [];
+    }
+
+    get propToValuesMap(): PropToValuesMap {
+        return this.cachedData ?
+            this.cachedData.propToValuesMap :
+            {};
+    }
+
+    get highlightingStats(): StationHighlightingStats | null {
+        return this.cachedData ?
+            this.cachedData.data.highlightingStats.stationRuleStats :
+            null;
+    }
 
     private isHighlightingStationTabActive$: Observable<boolean> = this.store.pipe(
         select(tracingSelectors.getIsHighlightingStationTabActive),
@@ -23,9 +69,13 @@ export class HighlightingStationComponent implements OnInit, OnDestroy {
 
     private componentIsActive = true;
     private stateSubscription: Subscription | null = null;
+    private cachedData: CachedData | null = null;
+    private cachedState: HighlightingState | null = null;
 
     constructor(
-        private store: Store<fromTracing.State>,
+        private editHighlightingService: EditHighlightingService,
+        private dataService: DataService,
+        private store: Store<tracingReducers.State>,
         private alertService: AlertService
     ) { }
 
@@ -40,15 +90,29 @@ export class HighlightingStationComponent implements OnInit, OnDestroy {
                     }
                 } else {
                     if (!this.stateSubscription) {
-                        this.stateSubscription = this.store.select(tracingSelectors.getStationHighlightingSettings).subscribe(
-                            (stationHighlightingData) => this.applyState(stationHighlightingData),
-                            err => this.alertService.error(`getStationFilterData store subscription failed: ${err}`)
+                        this.stateSubscription = this.store.select(tracingSelectors.getStationHighlightingData).subscribe(
+                            (state: HighlightingState) => this.applyState(state),
+                            err => this.alertService.error(`getStationHighlightingData store subscription failed: ${err}`)
                         );
                     }
                 }
             },
             err => this.alertService.error(`showConfigurationSideBar store subscription failed: ${err}`)
         );
+    }
+
+    onRulesChange(newRules: StationHighlightingRule[]) {
+        this.emitNewRules(newRules);
+    }
+
+    onColorOrShapeRuleEditIndexChange(editIndex: number | null) {
+        this.emitColorOrShapeRuleEditIndexChange(editIndex);
+    }
+
+    onRuleDelete(deleteRuleRequestData: HighlightingRuleDeleteRequestData) {
+        this.store.dispatch(new configurationActions.DeleteStationHighlightingRulesSSA(
+            { stationHighlightingRule: deleteRuleRequestData }
+        ));
     }
 
     ngOnDestroy() {
@@ -59,9 +123,50 @@ export class HighlightingStationComponent implements OnInit, OnDestroy {
         }
     }
 
-    private applyState(stationHighlightingData: StationHighlightingData[]) {
-        this.colorAndShapeHighlightings = stationHighlightingData.filter((item: StationHighlightingData) => {
-            return (item.color || (item.shape !== undefined && item.shape !== null));
-        });
+    private emitNewRules(rules: StationHighlightingRule[]) {
+        this.store.dispatch(new tracingActions.SetStationHighlightingRulesSOA(
+            { rules: rules }
+        ));
     }
+
+    private emitColorOrShapeRuleEditIndexChange(editIndex: number | null) {
+        this.store.dispatch(new tracingActions.SetColorsAndShapesEditIndexSOA(
+            { editIndex: editIndex }
+        ));
+    }
+
+    private applyState(state: HighlightingState): void {
+        let dataTable: DataTable | null = this.cachedData ? this.cachedData.dataTable : null;
+        const data = this.dataService.getData(state.graphState);
+        if (!this.cachedState || this.cachedState.graphState.fclElements !== state.graphState.fclElements) {
+            dataTable = this.editHighlightingService.getStationData(state.graphState);
+        } else if (
+            data.stations !== this.cachedData.data.stations ||
+            data.deliveries !== this.cachedData.data.deliveries ||
+            data.tracingResult !== this.cachedData.data.tracingResult ||
+            data.statSel !== this.cachedData.data.statSel ||
+            data.delSel !== this.cachedData.data.delSel
+            ) {
+            dataTable = {
+                ...this.editHighlightingService.getStationData(state.graphState),
+                columns: this.cachedData.dataTable.columns
+            };
+        }
+
+        this.cachedState = {
+            ...state
+        };
+
+        const propToValuesMap: PropToValuesMap =
+            this.cachedData === null || this.cachedData.dataTable !== dataTable ?
+            ComplexFilterUtils.extractPropToValuesMap(dataTable, dataTable.columns) :
+            this.cachedData.propToValuesMap;
+
+        this.cachedData = {
+            dataTable: dataTable,
+            propToValuesMap: propToValuesMap,
+            data: data
+        };
+    }
+
 }
