@@ -1,14 +1,19 @@
-import { FclData, ROASettings } from './data.model';
+import { FclData } from './data.model';
 import { VisioReport } from './visio/layout-engine/datatypes';
-import { ConfigurationTabIndex, FilterSettings } from './configuration/configuration.model';
+import { ConfigurationTabIndex, FilterSettings, HighlightingConfigurationSettings } from './configuration/configuration.model';
+import { ROASettings } from './visio/model';
 
-export interface TracingState {
-    fclData: FclData;
+export interface ModelDependentState {
     visioReport: VisioReport | null;
-    roaSettings: ROASettings;
+    roaSettings: ROASettings | null;
+    filterSettings: FilterSettings;
+    highlightingConfigurationSettings: HighlightingConfigurationSettings;
+}
+
+export interface TracingState extends ModelDependentState {
+    fclData: FclData;
     showGraphSettings: boolean;
     showConfigurationSideBar: boolean;
     configurationTabIndices: ConfigurationTabIndex;
-    filterSettings: FilterSettings;
     tracingActive: boolean;
 }
