@@ -1,5 +1,6 @@
 import { Component, OnInit, Input } from '@angular/core';
-import { LabelElementInfo } from '@app/tracing/data.model';
+import { AmountUnitPair, LabelElementInfo, PropElementInfo } from '../model';
+import { getUnitPropFromAmountProp } from '../shared';
 
 @Component({
     selector: 'fcl-label-configuration-view',
@@ -10,9 +11,23 @@ export class LabelConfigurationViewComponent implements OnInit {
 
     @Input() labelElements: LabelElementInfo[][];
     @Input() availableProps: { prop: string, label: string }[];
+    @Input() amountUnitPairs: AmountUnitPair[] = [];
 
     constructor() { }
 
     ngOnInit() {
+    }
+
+    onPropElementPropChange(propElement: PropElementInfo, prop: string | null): void {
+        propElement.prop = prop;
+        for (const pair of this.amountUnitPairs) {
+            if (pair.amount === propElement) {
+                pair.unit.prop = getUnitPropFromAmountProp(propElement.prop, this.availableProps);
+            }
+        }
+    }
+
+    onPropElementAltTextChange(propElement: PropElementInfo, altText: string): void {
+        propElement.altText = altText;
     }
 }
