@@ -1,4 +1,5 @@
-import { Component, OnInit, Output, Input, EventEmitter } from '@angular/core';
+import { Component, OnInit, Output, Input, EventEmitter, OnChanges, SimpleChanges } from '@angular/core';
+import { MatSelectChange } from '@angular/material/select';
 import { PropInfo } from '@app/tracing/shared/property-info';
 
 @Component({
@@ -6,16 +7,27 @@ import { PropInfo } from '@app/tracing/shared/property-info';
     templateUrl: './property-element-view.component.html',
     styleUrls: ['./property-element-view.component.scss']
 })
-export class PropertyElementViewComponent implements OnInit {
+export class PropertyElementViewComponent implements OnInit, OnChanges {
 
     @Input() availableProps: PropInfo[];
-    @Input() prop: string;
+    @Input() prop: string | null = null;
     @Input() altText: string;
+    @Input() enableNullProp = false;
 
-    @Output() propChange = new EventEmitter<string>();
+    @Output() propChange = new EventEmitter<string | null>();
     @Output() altTextChange = new EventEmitter<string>();
 
+    private propWrapper_: string = '';
+
+    get propWrapper(): string {
+        return this.propWrapper_;
+    }
+
     constructor() { }
+
+    ngOnChanges(changes: SimpleChanges): void {
+        this.propWrapper_ = this.prop || '';
+    }
 
     ngOnInit() {
     }
@@ -24,12 +36,11 @@ export class PropertyElementViewComponent implements OnInit {
         return this.availableProps.some((p) => p.prop === this.prop);
     }
 
-    setProperty(prop: string): void {
-        this.propChange.emit(this.prop);
-    }
-
     onAltTextChange(altText: string): void {
         this.altTextChange.emit(altText);
     }
 
+    onSelectionChange(change: MatSelectChange): void {
+        this.propChange.emit(change.value || null);
+    }
 }
